+++ test/datascript/test/pull_api.cljs
@@ -5,78 +5,77 @@
    [cemerick.cljs.test :as t]
    [datascript :as d]))
 
-(defn- create-test-db
-  []
-  (-> {:aka    { :db/cardinality :db.cardinality/many }
-       :child  { :db/cardinality :db.cardinality/many
-                :db/valueType :db.type/ref }
-       :friend { :db/cardinality :db.cardinality/many
-                :db/valueType :db.type/ref }
-       :enemy  { :db/cardinality :db.cardinality/many
-                :db/valueType :db.type/ref }
-       :father { :db/valueType :db.type/ref }
-       
-       :part   { :db/valueType :db.type/ref
-                :db/isComponent true
-                :db/cardinality :db.cardinality/many }}
-      d/empty-db
-      (d/db-with [[:db/add 1 :name  "Ivan"]
-                  [:db/add 1 :name  "Petr"]
-                  [:db/add 1 :aka   "Devil"]
-                  [:db/add 1 :aka   "Tupen"]
-                  [:db/add 2 :name  "David"]
-                  [:db/add 3 :name  "Thomas"]
-                  [:db/add 4 :name  "Lucy"]
-                  [:db/add 5 :name  "Elizabeth"]
-                  [:db/add 6 :name  "Matthew"]
-                  [:db/add 7 :name  "Eunan"]
-                  [:db/add 8 :name  "Kerri"]
-                  [:db/add 9 :name  "Rebecca"]
-                  [:db/add 1 :child 2]
-                  [:db/add 1 :child 3]
-                  [:db/add 6 :father 3]
-                  
-                  [:db/add 10 :name  "Part A"]
-                  [:db/add 11 :name  "Part A.A"]
-                  [:db/add 10 :part 11]
-                  [:db/add 12 :name  "Part A.A.A"]
-                  [:db/add 11 :part 12]
-                  [:db/add 13 :name  "Part A.A.A.A"]
-                  [:db/add 12 :part 13]
-                  [:db/add 14 :name  "Part A.A.A.B"]
-                  [:db/add 12 :part 14]
-                  [:db/add 15 :name  "Part A.B"]
-                  [:db/add 10 :part 15]
-                  [:db/add 16 :name  "Part A.B.A"]
-                  [:db/add 15 :part 16]
-                  [:db/add 17 :name  "Part A.B.A.A"]
-                  [:db/add 16 :part 17]
-                  [:db/add 18 :name  "Part A.B.A.B"]
-                  [:db/add 16 :part 18]])))
+(def ^:private test-schema
+  {:aka    { :db/cardinality :db.cardinality/many }
+   :child  { :db/cardinality :db.cardinality/many
+             :db/valueType :db.type/ref }
+   :friend { :db/cardinality :db.cardinality/many
+             :db/valueType :db.type/ref }
+   :enemy  { :db/cardinality :db.cardinality/many
+             :db/valueType :db.type/ref }
+   :father { :db/valueType :db.type/ref }
+
+   :part   { :db/valueType :db.type/ref
+             :db/isComponent true
+             :db/cardinality :db.cardinality/many }})
+
+(def test-datoms
+  (->>
+    [[1 :name  "Petr"]
+     [1 :aka   "Devil"]
+     [1 :aka   "Tupen"]
+     [2 :name  "David"]
+     [3 :name  "Thomas"]
+     [4 :name  "Lucy"]
+     [5 :name  "Elizabeth"]
+     [6 :name  "Matthew"]
+     [7 :name  "Eunan"]
+     [8 :name  "Kerri"]
+     [9 :name  "Rebecca"]
+     [1 :child 2]
+     [1 :child 3]
+     [6 :father 3]
+     [10 :name  "Part A"]
+     [11 :name  "Part A.A"]
+     [10 :part 11]
+     [12 :name  "Part A.A.A"]
+     [11 :part 12]
+     [13 :name  "Part A.A.A.A"]
+     [12 :part 13]
+     [14 :name  "Part A.A.A.B"]
+     [12 :part 14]
+     [15 :name  "Part A.B"]
+     [10 :part 15]
+     [16 :name  "Part A.B.A"]
+     [15 :part 16]
+     [17 :name  "Part A.B.A.A"]
+     [16 :part 17]
+     [18 :name  "Part A.B.A.B"]
+     [16 :part 18]]
+   (map #(apply d/datom %))))
+
+(def ^:private test-db (d/init-db test-datoms test-schema))
 
 (deftest test-pull-attr-spec
-  (let [db (create-test-db)]
-    (is (= {:name "Petr" :aka ["Devil" "Tupen"]}
-           (d/pull db '[:name :aka] 1)))
+  (is (= {:name "Petr" :aka ["Devil" "Tupen"]}
+         (d/pull test-db '[:name :aka] 1)))
 
-    (is (= {:name "Matthew" :father {:db/id 3} :db/id 6}
-           (d/pull db '[:name :father :db/id] 6)))
+  (is (= {:name "Matthew" :father {:db/id 3} :db/id 6}
+         (d/pull test-db '[:name :father :db/id] 6)))
 
-    (is (= [{:name "Petr"} {:name "Elizabeth"}
-            {:name "Eunan"} {:name "Rebecca"}]
-           (d/pull-many db '[:name] [1 5 7 9])))))
+  (is (= [{:name "Petr"} {:name "Elizabeth"}
+          {:name "Eunan"} {:name "Rebecca"}]
+         (d/pull-many test-db '[:name] [1 5 7 9]))))
 
 (deftest test-pull-reverse-attr-spec
-  (let [db (create-test-db)]
-    (is (= {:name "David" :_child [{:db/id 1}]}
-           (d/pull db '[:name :_child] 2)))
+  (is (= {:name "David" :_child [{:db/id 1}]}
+         (d/pull test-db '[:name :_child] 2)))
 
-    (is (= {:name "David" :_child [{:name "Petr"}]}
-           (d/pull db '[:name {:_child [:name]}] 2)))))
+  (is (= {:name "David" :_child [{:name "Petr"}]}
+         (d/pull test-db '[:name {:_child [:name]}] 2))))
 
 (deftest test-pull-component-attr
-  (let [db    (create-test-db)
-        parts {:name "Part A",
+  (let [parts {:name "Part A",
                :part
                [{:db/id 11
                  :name "Part A.A",
@@ -95,26 +94,28 @@
                    [{:db/id 17 :name "Part A.B.A.A"}
                     {:db/id 18 :name "Part A.B.A.B"}]}]}]}]
     (testing "Component entities are expanded recursively"
-      (is (= parts (d/pull db '[:name :part] 10))))
+      (is (= parts (d/pull test-db '[:name :part] 10))))
 
     (testing "Reverse component references yield a single result"
       (is (= {:name "Part A.A" :_part {:db/id 10}}
-             (d/pull db [:name :_part] 11))))))
+             (d/pull test-db [:name :_part] 11))))))
 
 (deftest test-pull-wildcard
-  (let [db (create-test-db)]
-    (is (= {:db/id 1 :name "Petr" :aka ["Devil" "Tupen"]
-            :child [{:db/id 2} {:db/id 3}]}
-           (d/pull db '[*] 1)))))
+  (is (= {:db/id 1 :name "Petr" :aka ["Devil" "Tupen"]
+          :child [{:db/id 2} {:db/id 3}]}
+         (d/pull test-db '[*] 1))))
 
 (deftest test-pull-limit
-  (let [db (-> (create-test-db)
-               (d/db-with [[:db/add 4 :friend 5]
-                           [:db/add 4 :friend 6]
-                           [:db/add 4 :friend 7]
-                           [:db/add 4 :friend 8]])
-               (d/db-with (for [idx (range 2000)]
-                            [:db/add 8 :aka (str "aka-" idx)])))]
+  (let [db (d/init-db
+             (concat
+               test-datoms
+               [(d/datom 4 :friend 5)
+                (d/datom 4 :friend 6)
+                (d/datom 4 :friend 7)
+                (d/datom 4 :friend 8)]
+               (for [idx (range 2000)]
+                 (d/datom 8 :aka (str "aka-" idx))))
+              test-schema)]
 
     (testing "Without an explicit limit, the default is 1000"
       (is (->> (d/pull db '[:aka] 8)
@@ -146,35 +147,33 @@
              (d/pull db '[:name {(limit :friend 2) [:name]}] 4))))))
 
 (deftest test-pull-default
-  (let [db (create-test-db)]
-    (testing "Empty results return nil"
-      (is (nil? (d/pull db '[:foo] 1))))
+  (testing "Empty results return nil"
+    (is (nil? (d/pull test-db '[:foo] 1))))
 
-    (testing "A default can be used to replace nil results"
-      (is (= {:foo "bar"}
-             (d/pull db '[(default :foo "bar")] 1))))))
+  (testing "A default can be used to replace nil results"
+    (is (= {:foo "bar"}
+           (d/pull test-db '[(default :foo "bar")] 1)))))
 
 (deftest test-pull-map
-  (let [db (create-test-db)]
-    (testing "Single attrs yield a map"
-      (is (= {:name "Matthew" :father {:name "Thomas"}}
-             (d/pull db '[:name {:father [:name]}] 6))))
+  (testing "Single attrs yield a map"
+    (is (= {:name "Matthew" :father {:name "Thomas"}}
+           (d/pull test-db '[:name {:father [:name]}] 6))))
 
-    (testing "Multi attrs yield a collection of maps"
-      (is (= {:name "Petr" :child [{:name "David"}
-                                   {:name "Thomas"}]}
-             (d/pull db '[:name {:child [:name]}] 1))))
+  (testing "Multi attrs yield a collection of maps"
+    (is (= {:name "Petr" :child [{:name "David"}
+                                 {:name "Thomas"}]}
+           (d/pull test-db '[:name {:child [:name]}] 1))))
 
-    (testing "Missing attrs are dropped"
-      (is (= {:name "Petr"}
-             (d/pull db '[:name {:father [:name]}] 1))))
+  (testing "Missing attrs are dropped"
+    (is (= {:name "Petr"}
+           (d/pull test-db '[:name {:father [:name]}] 1))))
 
-    (testing "Non matching results are removed from collections"
-      (is (= {:name "Petr" :child []}
-             (d/pull db '[:name {:child [:foo]}] 1))))))
+  (testing "Non matching results are removed from collections"
+    (is (= {:name "Petr" :child []}
+           (d/pull test-db '[:name {:child [:foo]}] 1)))))
 
 (deftest test-pull-recursion
-  (let [db      (-> (create-test-db)
+  (let [db      (-> test-db
                     (d/db-with [[:db/add 4 :friend 5]
                                 [:db/add 5 :friend 6]
                                 [:db/add 6 :friend 7]
@@ -233,24 +232,23 @@
                (d/pull db '[:db/id :name {:friend ...}] 4)))))))
 
 (deftest test-deep-recursion
-  (let [db    (create-test-db)
-        start 100
-        depth 10000
+  (let [start 100
+        depth 1500
         txd   (mapcat
                (fn [idx]
-                 [[:db/add idx :name (str "Person-" idx)]
-                  [:db/add (dec idx) :friend idx]])
+                 [(d/datom idx :name (str "Person-" idx))
+                  (d/datom (dec idx) :friend idx)])
                (range (inc start) depth))
-        db    (-> db
-                  (d/db-with [[:db/add start :name (str "Person-" start)]])
-                  (d/db-with txd))]
-
+        db    (d/init-db (concat
+                           test-datoms
+                           [(d/datom start :name (str "Person-" start))]
+                           txd)
+                         test-schema)
+        pulled (d/pull db '[:name {:friend ...}] start)
+        path   (->> [:friend 0]
+                    (repeat (dec (- depth start)))
+                    (into [] cat))]
     (is (= (str "Person-" (dec depth))
-           (->> (d/pull db '[:name {:friend ...}] start)
-                (iterate #(get-in % [:friend 0]))
-                (drop (dec (- depth start)))
-                first
-                :name)))))
+           (:name (get-in pulled path))))))
 
-
-#_(t/test-ns 'datascript.test.pull)
+#_(t/test-ns 'datascript.test.pull-api)